objclass, osd: dissect the OSD-specific parts from class_api.cc.
